+++ packages/components/src/Form/Inputs/InputDate/index.ts
@@ -24,6 +24,4 @@
 
  */
 
-export * from './IconList'
-export * from './DateTimeFormat'
-export * from './CalendarPropTable'
+export * from './InputDate'